+++ lib/downloader/t_progress_dialog.dart
@@ -56,11 +56,7 @@ class _TProgressDialogState extends State<TProgressDialog> {
       title: widget.title,
       content: TScrollableColumn(
         children: [
-          errorMsg == null
-              ? progress == null
-                    ? Text('Preparing...')
-                    : _getMessage()
-              : Text(errorMsg!, style: TextStyle(color: Colors.red)),
+          _getMessage(),
           // progress
           _getProgress(),
         ],
@@ -72,6 +68,9 @@ class _TProgressDialogState extends State<TProgressDialog> {
   Widget _getMessage() {
     if (progress == null) {
       return SizedBox.shrink();
+    }
+    if (errorMsg != null) {
+      return Text(errorMsg!, style: TextStyle(color: Colors.red));
     }
     return Text(progress!.message);
   }
@@ -88,7 +87,7 @@ class _TProgressDialogState extends State<TProgressDialog> {
       children: [
         // progress lable
         progress!.status == TProgressTypes.preparing
-            ? Text(progress!.status.name)
+            ? SizedBox.shrink()
             : Text('${progress!.index}/${progress!.indexLength}'),
         // progress
         LinearProgressIndicator(